+++ Source/WebKit/UIProcess/XR/openxr/PlatformXROpenXR.cpp
@@ -30,6 +30,10 @@
 
 #include "APIUIClient.h"
 #include "OpenXRExtensions.h"
+#include "OpenXRGraphicsBinding.h"
+#if defined(XR_USE_GRAPHICS_API_OPENGL_ES)
+#include "OpenXRGraphicsBindingOpenGLES.h"
+#endif
 #include "OpenXRHitTestManager.h"
 #include "OpenXRInput.h"
 #include "OpenXRInputSource.h"
@@ -37,23 +41,8 @@
 #include "OpenXRUtils.h"
 #include "WebPageProxy.h"
 #include "WebProcessProxy.h"
-#if USE(LIBEPOXY)
-#define __GBM__ 1
-#include <epoxy/egl.h>
-#else
-#include <EGL/egl.h>
-#endif
-#include <WebCore/GLContext.h>
-#include <WebCore/GLDisplay.h>
-#include <WebCore/GLFence.h>
 #include <wtf/RunLoop.h>
 #include <wtf/WorkQueue.h>
-
-#if USE(GBM)
-#include "DRMMainDevice.h"
-#include <WebCore/DRMDevice.h>
-#include <WebCore/GBMDevice.h>
-#endif
 
 #if OS(ANDROID)
 #include <dlfcn.h>
@@ -196,10 +185,7 @@ bool OpenXRCoordinator::collectSwapchainFormatsIfNeeded()
 
 std::unique_ptr<OpenXRSwapchain> OpenXRCoordinator::createSwapchain(uint32_t width, uint32_t height, bool alpha, uint32_t faceCount) const
 {
-    // Even if alpha is false we always ask for the RGBA8 format, as the DRM_FORMAT_RGB8 is not supported by ANGLE.
-    // In this case we ignore the alpha channel by using DRM_FORMAT_XRGB8888 when exporting the texture.
-    auto preferredFormat = GL_RGBA8;
-    auto format = m_supportedSwapchainFormats.contains(preferredFormat) ? preferredFormat : m_supportedSwapchainFormats.first();
+    auto format = m_graphicsBinding->selectColorFormat(m_supportedSwapchainFormats, alpha);
     auto sampleCount = m_viewConfigurationViews.isEmpty() ? 1 : m_viewConfigurationViews.first().recommendedSwapchainSampleCount;
 
     auto info = createOpenXRStruct<XrSwapchainCreateInfo, XR_TYPE_SWAPCHAIN_CREATE_INFO>();
@@ -212,7 +198,7 @@ std::unique_ptr<OpenXRSwapchain> OpenXRCoordinator::createSwapchain(uint32_t wid
     info.sampleCount = sampleCount;
     info.usageFlags = XR_SWAPCHAIN_USAGE_SAMPLED_BIT | XR_SWAPCHAIN_USAGE_COLOR_ATTACHMENT_BIT;
 
-    return OpenXRSwapchain::create(m_session, info, alpha ? OpenXRSwapchain::HasAlpha::Yes : OpenXRSwapchain::HasAlpha::No);
+    return OpenXRSwapchain::create(m_session, info, alpha ? OpenXRSwapchain::HasAlpha::Yes : OpenXRSwapchain::HasAlpha::No, *m_graphicsBinding);
 }
 
 void OpenXRCoordinator::createLayerProjection(uint32_t width, uint32_t height, bool alpha, CompletionHandler<void(std::optional<PlatformXR::LayerInfo>)>&& reply)
@@ -241,10 +227,6 @@ void OpenXRCoordinator::createLayerProjection(uint32_t width, uint32_t height, b
 
                 auto imageCount = swapchain->imageCount();
                 if (auto layer = OpenXRLayerProjection::create(WTF::move(swapchain))) {
-#if USE(GBM)
-                    if (m_gbmDevice)
-                        layer->setGBMDevice(m_gbmDevice);
-#endif
                     auto layerHandle = m_nextLayerHandle++;
                     m_layers.add(layerHandle, WTF::move(layer));
                     PlatformXR::LayerInfo layerInfo { layerHandle, imageCount };
@@ -345,10 +327,6 @@ void OpenXRCoordinator::createCompositionLayer(PlatformXR::CompositionLayerType 
                     return;
                 }
 
-#if USE(GBM)
-                if (m_gbmDevice)
-                    layer->setGBMDevice(m_gbmDevice);
-#endif
                 auto layerHandle = m_nextLayerHandle++;
                 m_layers.add(layerHandle, WTF::move(layer));
                 PlatformXR::LayerInfo layerInfo { layerHandle, imageCount };
@@ -675,13 +653,8 @@ void OpenXRCoordinator::createInstance()
 #endif
 
     Vector<char *> extensions;
-#if defined(XR_USE_PLATFORM_EGL)
-    if (OpenXRExtensions::singleton().isExtensionSupported(XR_MNDX_EGL_ENABLE_EXTENSION_NAME ""_span))
-        extensions.append(const_cast<char*>(XR_MNDX_EGL_ENABLE_EXTENSION_NAME));
-#endif
-#if defined(XR_USE_GRAPHICS_API_OPENGL_ES)
-    extensions.append(const_cast<char*>(XR_KHR_OPENGL_ES_ENABLE_EXTENSION_NAME));
-#endif
+    for (auto graphicsExtension : m_graphicsBinding->requiredInstanceExtensions())
+        extensions.append(const_cast<char*>(graphicsExtension.characters()));
 #if defined(XR_EXT_hand_interaction)
     if (OpenXRExtensions::singleton().isExtensionSupported(XR_EXT_HAND_INTERACTION_EXTENSION_NAME ""_span))
         extensions.append(const_cast<char*>(XR_EXT_HAND_INTERACTION_EXTENSION_NAME));
@@ -732,52 +705,6 @@ void OpenXRCoordinator::createInstance()
     CHECK_XRCMD(xrCreateInstance(&createInfo, &m_instance));
 }
 
-RefPtr<WebCore::GLDisplay> OpenXRCoordinator::createGLDisplay(bool isForTesting) const
-{
-    ASSERT(RunLoop::isMain());
-    ASSERT(!m_glDisplay);
-
-    const char* extensions = eglQueryString(nullptr, EGL_EXTENSIONS);
-    auto tryCreateDisplay = [&](EGLenum platform, void *native) -> RefPtr<WebCore::GLDisplay> {
-        if (WebCore::GLContext::isExtensionSupported(extensions, "EGL_EXT_platform_base"))
-            return WebCore::GLDisplay::create(eglGetPlatformDisplayEXT(platform, native, nullptr));
-        if (WebCore::GLContext::isExtensionSupported(extensions, "EGL_KHR_platform_base"))
-            return WebCore::GLDisplay::create(eglGetPlatformDisplay(platform, native, nullptr));
-        return nullptr;
-    };
-
-    RefPtr<WebCore::GLDisplay> glDisplay;
-
-#if OS(ANDROID)
-    if (WebCore::GLContext::isExtensionSupported(extensions, "EGL_KHR_platform_android")) {
-        glDisplay = tryCreateDisplay(EGL_PLATFORM_ANDROID_KHR, EGL_DEFAULT_DISPLAY);
-        if (!glDisplay)
-            glDisplay = WebCore::GLDisplay::create(eglGetDisplay(EGL_DEFAULT_DISPLAY));
-        if (glDisplay && !(glDisplay->extensions().ANDROID_get_native_client_buffer && glDisplay->extensions().ANDROID_image_native_buffer))
-            glDisplay = nullptr;
-    }
-#endif // OS(ANDROID)
-
-    if (WebCore::GLContext::isExtensionSupported(extensions, "EGL_MESA_platform_surfaceless")) {
-        glDisplay = tryCreateDisplay(EGL_PLATFORM_SURFACELESS_MESA, EGL_DEFAULT_DISPLAY);
-        if (glDisplay && !isForTesting && !glDisplay->extensions().MESA_image_dma_buf_export)
-            glDisplay = nullptr;
-    }
-
-#if USE(GBM)
-    if (!glDisplay && WebCore::GLContext::isExtensionSupported(extensions, "EGL_KHR_platform_gbm")) {
-        const auto& mainDevice = drmMainDevice();
-        if (!mainDevice.isNull()) {
-            m_gbmDevice = WebCore::GBMDevice::create(!mainDevice.renderNode.isNull() ? mainDevice.renderNode : mainDevice.primaryNode);
-            if (m_gbmDevice)
-                glDisplay = tryCreateDisplay(EGL_PLATFORM_GBM_KHR, m_gbmDevice->device());
-        }
-    }
-#endif
-
-    return glDisplay;
-}
-
 void OpenXRCoordinator::collectViewConfigurations()
 {
     ASSERT(RunLoop::isMain());
@@ -825,15 +752,22 @@ void OpenXRCoordinator::initializeDevice(bool isForTesting)
     if (m_instance != XR_NULL_HANDLE)
         return;
 
-    auto display = createGLDisplay(isForTesting);
-    if (!display) {
+    std::unique_ptr<OpenXRGraphicsBinding> graphicsBinding;
+#if defined(XR_USE_GRAPHICS_API_OPENGL_ES)
+    graphicsBinding = OpenXRGraphicsBindingOpenGLES::create();
+#endif
+    if (!graphicsBinding || !graphicsBinding->initializeDisplay(isForTesting)) {
         LOG(XR, "Failed to create a display for OpenXR.");
         return;
     }
+    m_graphicsBinding = WTF::move(graphicsBinding);
 
     createInstance();
     if (m_instance == XR_NULL_HANDLE) {
         LOG(XR, "Failed to create OpenXR instance.");
+        // The display is only kept once the instance has been created successfully, so that a
+        // later retry recreates the whole graphics binding from scratch.
+        m_graphicsBinding = nullptr;
         return;
     }
 
@@ -850,8 +784,6 @@ void OpenXRCoordinator::initializeDevice(bool isForTesting)
 
     collectViewConfigurations();
     initializeBlendModes();
-
-    m_glDisplay = WTF::move(display);
 }
 
 void OpenXRCoordinator::initializeBlendModes()
@@ -882,43 +814,6 @@ void OpenXRCoordinator::initializeBlendModes()
     m_vrBlendMode = supportsOpaqueBlendMode ? XR_ENVIRONMENT_BLEND_MODE_OPAQUE : m_arBlendMode;
 }
 
-void OpenXRCoordinator::tryInitializeGraphicsBinding()
-{
-#if !OS(ANDROID)
-    if (!OpenXRExtensions::singleton().isExtensionSupported(XR_MNDX_EGL_ENABLE_EXTENSION_NAME ""_span)) {
-        LOG(XR, "OpenXR MNDX_EGL_ENABLE extension is not supported.");
-        return;
-    }
-#endif
-
-    if (!m_glContext) {
-#if USE(GBM)
-        const WebCore::GLContext::Target target = m_gbmDevice ? WebCore::GLContext::Target::Default : WebCore::GLContext::Target::Surfaceless;
-#else
-        static const WebCore::GLContext::Target target = WebCore::GLContext::Target::Surfaceless;
-#endif
-        m_glContext = WebCore::GLContext::create(*m_glDisplay, target);
-        if (!m_glContext) {
-            LOG(XR, "Failed to create the GL context for OpenXR.");
-            return;
-        }
-        if (!m_glContext->makeContextCurrent()) {
-            LOG(XR, "Failed to make the GL context current.");
-            return;
-        }
-    }
-
-#if OS(ANDROID)
-    m_graphicsBinding = createOpenXRStruct<XrGraphicsBindingOpenGLESAndroidKHR, XR_TYPE_GRAPHICS_BINDING_OPENGL_ES_ANDROID_KHR>();
-#else
-    m_graphicsBinding = createOpenXRStruct<XrGraphicsBindingEGLMNDX, XR_TYPE_GRAPHICS_BINDING_EGL_MNDX>();
-    m_graphicsBinding.getProcAddress = OpenXRExtensions::singleton().methods().getProcAddressFunc;
-#endif
-    m_graphicsBinding.display = m_glDisplay->eglDisplay();
-    m_graphicsBinding.context = m_glContext->platformContext();
-    m_graphicsBinding.config = m_glContext->config();
-}
-
 void OpenXRCoordinator::createSessionIfNeeded()
 {
     ASSERT(!RunLoop::isMain());
@@ -927,19 +822,17 @@ void OpenXRCoordinator::createSessionIfNeeded()
     if (m_session != XR_NULL_HANDLE)
         return;
 
-#if defined(XR_USE_GRAPHICS_API_OPENGL_ES)
-    auto requirements = createOpenXRStruct<XrGraphicsRequirementsOpenGLESKHR, XR_TYPE_GRAPHICS_REQUIREMENTS_OPENGL_ES_KHR>();
-    CHECK_XRCMD(OpenXRExtensions::singleton().methods().xrGetOpenGLESGraphicsRequirementsKHR(m_instance, m_systemId, &requirements));
-#endif
-
-    tryInitializeGraphicsBinding();
+    if (!m_graphicsBinding->initializeForSession(m_instance, m_systemId)) {
+        LOG(XR, "Failed to initialize the graphics binding for the OpenXR session.");
+        return;
+    }
 
     m_views.resize(m_viewConfigurationViews.size());
 
     // Create the session.
     auto sessionCreateInfo = createOpenXRStruct<XrSessionCreateInfo, XR_TYPE_SESSION_CREATE_INFO>();
     sessionCreateInfo.systemId = m_systemId;
-    sessionCreateInfo.next = &m_graphicsBinding;
+    sessionCreateInfo.next = m_graphicsBinding->sessionGraphicsBinding();
     CHECK_XRCMD(xrCreateSession(m_instance, &sessionCreateInfo, &m_session));
 }
 
@@ -974,7 +867,8 @@ void OpenXRCoordinator::cleanupSessionAndAssociatedResources()
         m_session = XR_NULL_HANDLE;
     }
 
-    m_glContext.reset();
+    if (m_graphicsBinding)
+        m_graphicsBinding->releaseSessionGraphics();
 }
 
 void OpenXRCoordinator::cleanupInstanceAndAssociatedResources()
@@ -982,8 +876,7 @@ void OpenXRCoordinator::cleanupInstanceAndAssociatedResources()
     m_viewConfigurationViews.clear();
     m_systemId = XR_NULL_SYSTEM_ID;
 
-    ASSERT(!m_glContext);
-    m_glDisplay = nullptr;
+    m_graphicsBinding = nullptr;
 
     if (m_instance == XR_NULL_HANDLE)
         return;
@@ -1160,7 +1053,7 @@ PlatformXR::FrameData OpenXRCoordinator::populateFrameData(Box<RenderState> rend
     for (auto& layer : m_layers) {
         if (!renderState->activeLayerHandles.contains(layer.key))
             continue;
-        auto layerData = layer.value->startFrame();
+        auto layerData = layer.value->startFrame(*m_graphicsBinding);
         if (layerData) {
             auto layerDataRef = makeUniqueRef<PlatformXR::FrameData::LayerData>(WTF::move(*layerData));
             frameData.layers.add(layer.key, WTF::move(layerDataRef));
@@ -1323,12 +1216,10 @@ void OpenXRCoordinator::endFrame(Box<RenderState> renderState, Vector<PlatformXR
             continue;
         }
 
-        if (layer.fenceFD) {
-            if (auto fence = WebCore::GLFence::importFD(*m_glDisplay, WTF::move(layer.fenceFD)))
-                fence->serverWait();
-        }
+        if (layer.fenceFD)
+            m_graphicsBinding->waitFrameFence(WTF::move(layer.fenceFD));
 
-        auto headers = it->value->endFrame(layer, m_localSpace, m_views);
+        auto headers = it->value->endFrame(*m_graphicsBinding, layer, m_localSpace, m_views);
         frameEndLayers.appendVector(WTF::move(headers));
     }
